Protocolo de comunicação FINALIZADO e testado.

+++ src/usp/wirelezzgame/server/ServerMessageCallback.java
@@ -3,9 +3,9 @@ package usp.wirelezzgame.server;
 public interface ServerMessageCallback {
 	
 	void dadosJogador(String nomeJogador, String nomeCompleto, String facebookID);
-	void timeJogador(Integer idTime);
-	void interagirArea(Integer idArea, Integer latitude, Integer longitude, Integer acao);
-	void responderCaptcha(Integer idCaptcha, String resposta);
+	void timeJogador(int idTime);
+	void interagirArea(int idArea, double latitude, double longitude, int acao);
+	void responderCaptcha(int idCaptcha, String resposta);
 	void mensagemChatTodos(String mensagem);
 	void mensagemChatTime(String mensagem);
